fix dumb warning for dumb thing

+++ include/genie/util/Utility.h
@@ -46,7 +46,7 @@
 namespace genie {
 namespace util {
 
-static_assert(std::numeric_limits<int>::max() >= 2147483647ull && "We assume a sane int size in bounds checking");
+static_assert((unsigned long long)std::numeric_limits<int>::max() >= 2147483647ull && "We assume a sane int size in bounds checking");
 
 inline bool floatsEquals(const float a, const float b)
 {
